requirements.txt
asgiref==3.6.0
braintree==4.20.0
certifi==2023.5.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
filelock==3.12.0
idna==3.4
Pillow==9.5.0
pipenv==2023.5.19
platformdirs==3.5.1
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.23.0
virtualenv-clone==0.5.7
